* hans/crypto/valgrind/OTP-15644:
  crypto: Remove compiler warning on an fprintf in the otp test engine
  crypto: Fix valgrind error in api_ng.c ng_crypto_one_time for OpenSSL<1.0.1

+++ lib/crypto/c_src/otp_test_engine.c
@@ -160,7 +160,7 @@ static int test_engine_md5_update(EVP_MD_CTX *ctx,const void *data, size_t count
 
 static int test_engine_md5_final(EVP_MD_CTX *ctx,unsigned char *md) {
 #ifdef OLD
-    fprintf(stderr, "MD5 final size of EVP_MD: %lu\r\n", sizeof(EVP_MD));
+    fprintf(stderr, "MD5 final size of EVP_MD: %lu\r\n", (unsigned long)sizeof(EVP_MD));
     if (!MD5_Final(md, data(ctx)))
         goto err;
